==> VMS-datacall/hackathon_workflow/thuenen/0_global_Thuenen_01.R <==
#-------------------------------------------------------------------------------
#
# Script to extract and process VMS and logbook data for ICES VMS data call
#
# By: Niels Hintzen, Katell Hamon, Marcel Machiels#
# Code by: Niels Hintzen
# Contact: niels.hintzen@wur.nl
#
# Date: 25-Jan-2017
# Update Date: 29-Jan-2019 ; Updated by: Roi Martinez
# Update Date: 04-Feb-2020 ; Updated by: Colin Millar
# Update Date: 07-Feb 2020 ; Updated by: Neil Campbell
# Client: ICES
#-------------------------------------------------------------------------------

#--------------------READ ME----------------------------------------------------
# The following script is a proposed workflow example to processes the ICES
# VMS datacall request. It is not an exact template to be applied to data from
# every member state and needs to be adjusted according to the data availability
# and needs of every member state.
#-------------------------------------------------------------------------------


#- Clear workspace
rm(list=ls())

library(vmstools) #- download from www.vmstools.org
library(Matrix)   #- available on CRAN
library(ggplot2)  #- available on CRAN
library(dplyr)    #- available on CRAN
library(sp)
library(doBy)
library(mixtools)
library(tidyr)
library(glue)
library(gt)

 
#- Settings paths
# codePath  <- "VMSdatacall/R"          #Location where you store R scripts
# dataPath  <- "VMSdatacall/Data"       #Location where you store tacsat (VMS) and eflalo (logbook) data
# outPath   <- "VMSdatacall/Results"    #Location where you want to store the results


#- Settings paths DEU THUENEN
codePath   <- "/media/fssf01bhv/R/03_LinuxProjects/wg_WGSFD/VMS-datacall"                 #Location where you store R scripts
dataPath   <- "/media/fssf01bhv/vmsdata/eflalo_vms/0_IntermediateProducts/v202201"       #Location where you store tacsat (VMS) and eflalo (logbook) data
outPath    <- "/home/VTI_AD/schulze_t/R/03_LinuxProjects/wg_WGSFD_resultsThuenen/"               #Location where you want to store the results


#- Setting specific thresholds
spThres       <- 20   #Maximum speed threshold in analyses in nm
intThres      <- 5    #Minimum difference in time interval in minutes to prevent pseudo duplicates
intvThres     <- 240  #Maximum difference in time interval in minutes to prevent intervals being too large to be realistic
lanThres      <- 1.5  #Maximum difference in log10-transformed sorted weights

#- Re-run all years as we have new field for no. vessels
yearsToSubmit <- sort(c(2009,2016,2018,2020))

#- Set the gear names for which automatic fishing activity is wanted
#  It is important to fill out the gears you want to apply auto detection for
autoDetectionGears        <- c("TBB","OTB","OTT","SSC","SDN","DRB","PTB","HMD")

#- Decide if you want to visualy analyse speed-histograms to identify fishing activity
#  peaks or have prior knowledge and use the template provided around lines 380 below
visualInspection          <- FALSE

#- Specify how landings should be distributed over the VMS pings: By day, ICES rectangle, trip basis or otherwise
# linkEflaloTacsat          <- c("day","ICESrectangle","trip") # conserve=TRUE
# other options
# linkEflaloTacsat          <- c("day","ICESrectangle","trip") # conserve=TRUE
# linkEflaloTacsat          <- c("ICESrectangle","trip")       # conserve=TRUE
# linkEflaloTacsat          <- c("day","trip")                 # conserve=TRUE
linkEflaloTacsat          <- c("trip")                       # conserve=TRUE (trip and conserve=TRUE is not an option in splitAmongPings)
